# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Not My Road - Civic tech platform for reporting road issues in Delhi
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
}

export interface TimelineEvent {
  status: string;
  note?: string;
  timestamp: string;
}

export type ReportIssueType =
  (typeof ReportIssueType)[keyof typeof ReportIssueType];

export const ReportIssueType = {
  pothole: "pothole",
  broken_road: "broken_road",
  waterlogging: "waterlogging",
  damaged_divider: "damaged_divider",
  missing_signage: "missing_signage",
  other: "other",
} as const;

export type ReportRoadType =
  (typeof ReportRoadType)[keyof typeof ReportRoadType];

export const ReportRoadType = {
  main_road: "main_road",
  internal_road: "internal_road",
  highway: "highway",
} as const;

export type ReportAuthority =
  (typeof ReportAuthority)[keyof typeof ReportAuthority];

export const ReportAuthority = {
  MCD: "MCD",
  PWD: "PWD",
  NDMC: "NDMC",
  NHAI: "NHAI",
  DDA: "DDA",
  other: "other",
} as const;

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus];

export const ReportStatus = {
  pending: "pending",
  in_progress: "in_progress",
  resolved: "resolved",
} as const;

export interface Report {
  id: number;
  userId: number;
  userName: string;
  imageUrl: string;
  latitude: number;
  longitude: number;
  area: string;
  issueType: ReportIssueType;
  roadType: ReportRoadType;
  description?: string;
  authority: ReportAuthority;
  status: ReportStatus;
  tweetDraft: string;
  timeline: TimelineEvent[];
  daysUnresolved: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateReportInputIssueType =
  (typeof CreateReportInputIssueType)[keyof typeof CreateReportInputIssueType];

export const CreateReportInputIssueType = {
  pothole: "pothole",
  broken_road: "broken_road",
  waterlogging: "waterlogging",
  damaged_divider: "damaged_divider",
  missing_signage: "missing_signage",
  other: "other",
} as const;

export type CreateReportInputRoadType =
  (typeof CreateReportInputRoadType)[keyof typeof CreateReportInputRoadType];

export const CreateReportInputRoadType = {
  main_road: "main_road",
  internal_road: "internal_road",
  highway: "highway",
} as const;

export type CreateReportInputAuthority =
  (typeof CreateReportInputAuthority)[keyof typeof CreateReportInputAuthority];

export const CreateReportInputAuthority = {
  MCD: "MCD",
  PWD: "PWD",
  NDMC: "NDMC",
  NHAI: "NHAI",
  DDA: "DDA",
  other: "other",
} as const;

export interface CreateReportInput {
  userId: number;
  userName: string;
  imageUrl: string;
  latitude: number;
  longitude: number;
  area: string;
  issueType: CreateReportInputIssueType;
  roadType: CreateReportInputRoadType;
  description?: string;
  authority: CreateReportInputAuthority;
}

export type UpdateReportStatusInputStatus =
  (typeof UpdateReportStatusInputStatus)[keyof typeof UpdateReportStatusInputStatus];

export const UpdateReportStatusInputStatus = {
  pending: "pending",
  in_progress: "in_progress",
  resolved: "resolved",
} as const;

export interface UpdateReportStatusInput {
  status: UpdateReportStatusInputStatus;
  note?: string;
}

export interface Stats {
  totalReports: number;
  resolvedReports: number;
  resolvedPercentage: number;
  pendingReports: number;
  inProgressReports: number;
}

export type ListReportsParams = {
  /**
   * Filter reports by user ID
   */
  userId?: string;
  /**
   * Filter by status
   */
  status?: ListReportsStatus;
  limit?: number;
};

export type ListReportsStatus =
  (typeof ListReportsStatus)[keyof typeof ListReportsStatus];

export const ListReportsStatus = {
  pending: "pending",
  in_progress: "in_progress",
  resolved: "resolved",
} as const;

export type Logout200 = {
  message: string;
};