# lib/api-zod/src/generated/types/timelineEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Not My Road - Civic tech platform for reporting road issues in Delhi
 * OpenAPI spec version: 0.1.0
 */

export interface TimelineEvent {
  status: string;
  note?: string;
  timestamp: Date;
}